Separate out oracle utils
