+++ src/theory/quantifiers/oracle_engine.cpp
@@ -23,82 +23,16 @@
 #include "theory/quantifiers/first_order_model.h"
 #include "theory/quantifiers/term_registry.h"
 #include "theory/quantifiers/term_tuple_enumerator.h"
-#include "util/run.h"
+
 
 #include <stdlib.h>
-#include <iostream>
 
 using namespace cvc5::kind;
 using namespace cvc5::context;
 
-bool is_digits(const std::string &str)
-{
-    return str.find_first_not_of("0123456789") == std::string::npos;
-}
-
 namespace cvc5 {
 namespace theory {
 namespace quantifiers {
-
-Node OracleEngine::get_hex_numeral(std::string in)
-{
-  // we accept any sequence of '0'-'9', 'a'-'f', 'A'-'F'
-  std::size_t width = in.size()*16;
-  NodeManager* nm = NodeManager::currentNM();
-  unsigned int val = std::stoi(in, nullptr, 16);
-  Node result =  nm->mkConst(BitVector(width,val));
-  return result;
-}
-
-Node OracleEngine::get_bin_numeral(std::string in)
-{
-  // we accept any sequence of '0'-'1'
-  std::size_t width = in.size();
-  NodeManager* nm = NodeManager::currentNM();
-  unsigned int val = std::stoi(in, nullptr, 2);
-  Node result = nm->mkConst(BitVector(width,val));
-  return result;
-}
-
-Node OracleEngine::get_dec_numeral(std::string in)
-{
-  // we accept any sequence of '0'-'9'
-  NodeManager* nm = NodeManager::currentNM();
-  unsigned int val = std::stoi(in, nullptr, 10);
-  Node result  = nm->mkConst(Rational(val));
-  return result;
-}
-
-Node OracleEngine::responseParser(std::string &in)
-{
-  // Assumes the response is a singular integer or bitvector literal
-  // Temporary: will eventually be replaced with some subcomponent of full parser
-  NodeManager* nm = NodeManager::currentNM();
-  if(in.at(0)=='#')
-  {
-    if(in.at(1)=='b')
-      return get_bin_numeral(in);
-    else if(in.at(1)=='x')
-      return get_hex_numeral(in);
-    else
-      Assert(0); // throw error here
-  }
-  else if(is_digits(in))
-    return get_dec_numeral(in);
-  else if(in.find("true")!=std::string::npos)
-  {
-    Node result = nm->mkConst(true);
-    return result;
-  }
-  else if(in.find("false")!=std::string::npos)
-  {
-    Node result = nm->mkConst(false);
-    return result;
-  }
-  else
-    Assert(0); // throw error here
-}
-
 
 /** Attribute true for input variables */
 struct OracleInputVarAttributeId
@@ -132,62 +66,10 @@ void OracleEngine::reset_round(Theory::Effort e) {}
 
 void OracleEngine::registerQuantifier(Node q) {}
 
-std::string OracleEngine::getBinaryName(const Node n)
-{
-  // oracle functions have no children
-  if(n.getNumChildren()<3)
-    return n.getAttribute(OracleInterfaceAttribute());
-
-  // oracle interfaces have children, and the attribute is stored in 2nd child 
-  for (const Node& v : n[2][0]) 
-  { 
-    if(v.getAttribute(OracleInterfaceAttribute())!="")
-      return v.getAttribute(OracleInterfaceAttribute());
-  }
-  return "";
-} 
-
-
-
-Node OracleEngine::callOracle(const std::string &binary_name, 
-                                     const std::vector<Node> &argv)
-{
-  Trace("oracle-calls") << "Running oracle: " << binary_name ;
-  std::vector<std::string> string_args;
-  for (auto &arg : argv)
-  {
-    std::ostringstream oss;
-    oss << arg;
-    string_args.push_back(oss.str());
-    Trace("oracle-calls") << ' ' << arg;
-  }
-  Trace("oracle-calls") << std::endl;
-
-  // run the oracle binary
-  std::ostringstream stdout_stream;
-
-  auto run_result = run(
-      binary_name,
-      string_args,
-      "",
-      stdout_stream,
-      "");
-
-  // we assume that an oracle has a return code of 0 or 10. 
-  if (run_result != 0 && run_result !=10)
-  {
-    Trace("oracle-calls") << "oracle " << binary_name << " has failed with exit code " << run_result << std::endl;
-    Assert(run_result==0 || run_result==10);
-  }
-  // we assume that the oracle returns the result in SMT-LIB format
-  std::string stringResponse = stdout_stream.str();
-  // parse response into a Node
-  return responseParser(stringResponse);
-}
 
 void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
   double clSet = 0;
-  checkedAllOracles=false;
+  d_checkedAllOracles=false;
   if (Trace.isOn("oracle-engine"))
   {
     clSet = double(clock()) / double(CLOCKS_PER_SEC);
@@ -207,9 +89,9 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
       continue;
     }
     currInterfaces.push_back(q);
-    Trace("oracle-engine-state") << "Interface: " << q << " with binary name " << getBinaryName(q) << std::endl;
+    Trace("oracle-engine-state") << "Interface: " << q << " with binary name " << oracleCaller.getBinaryName(q) << std::endl;
   }
-  bool all_fapps_consistent=true;
+  bool allFappsConsistent=true;
   std::vector<Node> learned_lemmas;
   // iterate over oracle functions
   for (const Node& f : d_oracleFuns)
@@ -218,7 +100,7 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
     if(tat)
     {
       std::vector<Node> apps = tat->getLeaves(1);
-      std::string binaryName = getBinaryName(f);
+      std::string binaryName = oracleCaller.getBinaryName(f);
       Trace("oracle-calls") << "Oracle fun "<< f <<" with binary name "<< binaryName 
         <<" and " << apps.size()<< " applications."<< std::endl;
   
@@ -239,7 +121,7 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
         }
 
         // call oracle
-        Node response = callOracle(binaryName, arguments);  
+        Node response = oracleCaller.callOracle(binaryName, arguments);  
         Trace("oracle-calls") << "Node Response " << response;
         NodeManager* nm = NodeManager::currentNM();
         // check consistency with model
@@ -249,7 +131,7 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
         if(predictedResult!=response)
         {
           Trace("oracle-calls") << "Inconsistent response! Model expected " << predictedResult << std::endl;
-          all_fapps_consistent=false;
+          allFappsConsistent=false;
         }
         // add lemma
         Node fapp_with_values = nm->mkNode(APPLY_UF, arguments);
@@ -259,10 +141,10 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
     }
   }
   // if all were consistent, we can terminate
-  if(all_fapps_consistent)
+  if(allFappsConsistent)
   {
     Trace("oracle-engine-state") << "All responses consistent, no lemmas added" << std::endl;
-    consistencyCheckPassed=true;
+    d_consistencyCheckPassed=true;
   }
   else
   {
@@ -272,7 +154,7 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
       d_qim.lemma(l, InferenceId::QUANTIFIERS_ORACLE_INTERFACE);
     }
   }  
-  checkedAllOracles=true;
+  d_checkedAllOracles=true;
   // general SMTO: call constraint generators and assumption generators here
   
   if (Trace.isOn("oracle-engine"))
@@ -286,11 +168,11 @@ void OracleEngine::check(Theory::Effort e, QEffort quant_e) {
 bool OracleEngine::checkCompleteFor(Node q)
 {
   // TODO: true if oracle consistency check works
-  if(consistencyCheckPassed)
+  if(d_consistencyCheckPassed)
     Trace("oracle-engine-state") << q << " is complete"<< std::endl;
   else
     Trace("oracle-engine-state") << q << " is incomplete"<< std::endl;
-  return consistencyCheckPassed;
+  return d_consistencyCheckPassed;
 }
 
 void OracleEngine::checkOwnership(Node q)
