+++ src/util/CMakeLists.txt
@@ -43,6 +43,8 @@ libcvc5_add_sources(
   index.cpp
   index.h
   maybe.h
+  oracle_caller.cpp
+  oracle_caller.h
   ostream_util.cpp
   ostream_util.h
   poly_util.cpp
